[LinkControl] Extract reusable parts (#23869)
* Slice-and-dice LinkControl to make parts of it reusable

* Remove out of scope changes to search create button

* Remove out of scope changes

* Add missing files

* Simplify handleSelectSuggestion

* Update LinkControlSearchResults props description

* Update snapshots

* Adjust directEntry handler usage to be consistent with the master branch

* Add mock

* Restore the intended order of results in useSearchHandler

* Use state in Creating Entities test

* Fix two other createSuggestion tests

* Fix all unit tests

* Remove the dependency on saveEntityRecord

* Fix unit tests

* Rename LinkOverview to LinkPreview

* Remove console.log
