+++ packages/block-editor/src/components/link-control/constants.js
@@ -1,0 +1,4 @@
+// Used as a unique identifier for the "Create" option within search results.
+// Used to help distinguish the "Create" suggestion within the search results in
+// order to handle it as a unique case.
+export const CREATE_TYPE = '__CREATE__';
