+++ packages/block-editor/src/components/link-control/viewer-slot.js
@@ -1,0 +1,11 @@
+/**
+ * WordPress dependencies
+ */
+import { createSlotFill } from '@wordpress/components';
+
+const { Slot: ViewerSlot, Fill: ViewerFill } = createSlotFill(
+	'BlockEditorLinkControlViewer'
+);
+
+export { ViewerSlot, ViewerFill };
+export default ViewerSlot;
